Update BasePopulationStrategy.java
Steve, this is the change I mentioned in my email  yesterday that worked for me in order to get a full history of renamed files, but I started with the code as of 9/20, so I had to merge about a half dozen changes in and I haven't really tested it as such.  I couldn't get the current version to work with the approach of creating view labels for version 1 and the latest version of every source file in order to get a full history of every starteam change, so I kept the 9/20 version.